+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `config.preferSemicolons` which can have a Boolean value to determine whether the example code in the "Usage" section should be terminated by semicolons.
 - Added `config.quoteType` which can have a value of "single" or "double" to determine the type of quotes used in the "Usage" section.
 - Added an "API" section to the readme.
+- Added `config.sectionOverrides.{install,usage,testing,license}` which can have a string value to override the corresponding sections body text.
 
 ### Changed
 - The "Usage" section output is now determined by `config.pkg.type`. If it is defined as "module" an `import` statement is output, otherwise a `require` statement is output.